+++ conn_test.go
@@ -2,9 +2,6 @@ package gws
 
 import (
 	"bytes"
-	"compress/flate"
-	"encoding/binary"
-	"io"
 	"net"
 	"sync"
 	"sync/atomic"
@@ -54,68 +51,6 @@ func (c *webSocketMocker) OnPong(socket *Conn, payload []byte) {
 func (c *webSocketMocker) OnMessage(socket *Conn, message *Message) {
 	if c.onMessage != nil {
 		c.onMessage(socket, message)
-	}
-}
-
-func BenchmarkCompress(b *testing.B) {
-	var s = internal.AlphabetNumeric.Generate(1024)
-	var buf = bytes.NewBuffer(nil)
-	fw, _ := flate.NewWriter(nil, -2)
-	for i := 0; i < b.N; i++ {
-		buf.Reset()
-		fw.Reset(buf)
-		fw.Write(s)
-		fw.Flush()
-		fw.Close()
-	}
-}
-
-func BenchmarkDeCompress(b *testing.B) {
-	var s = internal.AlphabetNumeric.Generate(1024)
-	var buf = bytes.NewBuffer(nil)
-	fw, _ := flate.NewWriter(buf, -2)
-	fw.Write(s)
-	fw.Flush()
-	fw.Close()
-	fr := flate.NewReader(nil)
-
-	var result = bytes.NewBuffer(nil)
-	result.Grow(1024)
-	for i := 0; i < b.N; i++ {
-		r := bytes.NewBuffer(buf.Bytes())
-		fr.(flate.Resetter).Reset(r, nil)
-		result.Reset()
-		io.Copy(result, fr)
-	}
-}
-
-func BenchmarkMask(b *testing.B) {
-	var s1 = internal.AlphabetNumeric.Generate(1280)
-	var s2 = s1
-	var key [4]byte
-	binary.LittleEndian.PutUint32(key[:4], internal.AlphabetNumeric.Uint32())
-	for i := 0; i < b.N; i++ {
-		internal.MaskXOR(s2, key[:4])
-	}
-}
-
-func BenchmarkSyncMap_Store(b *testing.B) {
-	var m = sync.Map{}
-	for i := 0; i < b.N; i++ {
-		var key = string(internal.AlphabetNumeric.Generate(16))
-		m.Store(key, 1)
-	}
-}
-
-func BenchmarkSyncMap_Load(b *testing.B) {
-	var m = sync.Map{}
-	for i := 0; i < 10000; i++ {
-		var key = string(internal.AlphabetNumeric.Generate(4))
-		m.Store(key, 1)
-	}
-	for i := 0; i < b.N; i++ {
-		var key = string(internal.AlphabetNumeric.Generate(4))
-		m.Load(key)
 	}
 }
